apps/stig/pyproject.toml
[project]
name = "netnynja-stig"
version = "0.1.0"
description = "NetNynja Enterprise - STIG Manager Service"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "NetNynja Team" }
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.25.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.26.0",
    "nats-py>=2.6.0",
    "structlog>=24.1.0",
    # SSH/Device connectivity
    "asyncssh>=2.14.0",
    "netmiko>=4.3.0",
    "paramiko>=3.4.0",
    # Report generation
    "reportlab>=4.0.0",
    "weasyprint>=60.0",
    "jinja2>=3.1.0",
    # XML parsing for XCCDF/CKL
    "lxml>=5.0.0",
    "defusedxml>=0.7.0",
    # Observability
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-instrumentation-asyncpg>=0.43b0",
    "opentelemetry-exporter-otlp>=1.22.0",
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "black>=23.12.0",
    "mypy>=1.8.0",
    "types-redis>=4.6.0",
    "types-paramiko>=3.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stig"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
